1.2.3 fixes
- updated podspec
- added open / close basement methods (thx
https://github.com/bomjkolyadun)
- prepareForReuse fix (thx https://github.com/bomjkolyadun)
- fixed warnings for arm64
- demo app shows the basement on startup to hint the user (thx
https://github.com/bomjkolyadun)
- updated README

+++ BMXSwipableCellDemo/BMXDataItem.m
@@ -21,7 +21,7 @@
 
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"(%d) %@/%@", _userId, _firstName, _lastName];
+    return [NSString stringWithFormat:@"(%lu) %@/%@", (unsigned long)_userId, _firstName, _lastName];
 }
 
 @end
